third_party/WebKit/LayoutTests/html/dialog/multiple-centered-dialogs-expected.txt

Test that multiple dialogs are centered properly. You should see four boxes: one on the top row, two on the middle, and one on the bottom.
showing top-dialog
PASS dialog.offsetTop is expectedTop(dialog)
showing first-middle-dialog
PASS dialog.offsetTop is expectedTop(dialog)
showing second-middle-dialog
PASS dialog.offsetTop is expectedTop(dialog)
showing bottom-dialog
PASS dialog.offsetTop is expectedTop(dialog)
PASS successfullyParsed is true

TEST COMPLETE
